split generic BLE transport out of the Wendy protocol layer
internal/cli/ble mixed a reusable BLE central with Wendy's protocol
clients. Its own README already flagged the seam, down to naming the one
place it leaked. This makes the split physical, so the compiler enforces
it instead of a convention:

  ble/central/  generic transport — Connection (GATT + L2CAP) and the
                L2CAP-to-net.Conn adapter TLS runs over
  ble/scan/     generic scanning (already present)
  ble/          the Wendy protocol clients, and only those

conn.go moves to central/ minus the two Wendy pieces it carried:
NewClientTLSConfig, which depends on the Wendy PKI in shared/certs, and
DefaultL2CAPPSM. Neither subpackage now imports shared/models or
shared/certs, which is the property worth keeping.

timeoutSeconds is exported as central.TimeoutSeconds because
ble/lite_info.go needs the same round-up-never-below-1 behaviour, and
reimplementing that is a subtle bug waiting to happen.

Also splits DefaultL2CAPPSM in two. One constant served both the agent
and Wendy Lite, but those are two different programs' listening PSMs —
one an agent we ship, the other ESP32 firmware from another repo — so
changing either silently changed the other. Each protocol now owns its
own: ble.DefaultL2CAPPSM in agent_client.go, liteclient.DefaultL2CAPPSM
in link_ble.go. Both remain 128; no behaviour change.

Two doc comments in lite_info.go needed care: they describe a Lite
device but live in package ble, where a bare DefaultL2CAPPSM now names
the agent's constant. They are qualified explicitly.

Verified: go build, vet, gofmt, tests, and the GOOS=windows
cross-compile gate all clean; darwin builds both cgo bridges.
