+++ go/internal/cli/ble/central/ble_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package ble
+package central
 
 import "fmt"
